+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added Power method to Calculator class
+
 ## [0.1.1] - 2025-09-08
 
 ### Added
@@ -24,6 +28,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added XML documentation comments to test methods and constructor for improved code maintainability
 
 [unreleased]: https://github.com/neolution-ch/Neolution.ReleaseProcessTest.Test2/compare/v0.1.1...HEAD
-
 [0.1.1]: https://github.com/neolution-ch/Neolution.ReleaseProcessTest.Test2/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/neolution-ch/Neolution.ReleaseProcessTest.Test2/compare/v0.1.0-rc.0...v0.1.0